requirements.txt
beautifulsoup4==4.6.0
bleach==2.1.3
brotlipy==0.7.0
certifi==2018.4.16
cffi==1.11.5
csscompressor==0.9.5
Django==1.11.3
django-appconf==1.0.2
django-compressor==2.2
django-disqus==0.5
django-froala-editor==2.7.6.post2
django-htmlmin==0.10.0
django-imagekit==4.0.2
html5lib==1.0.1
lxml==4.1.1
mock==2.0.0
pbr==3.1.1
pilkit==2.0
Pillow==5.1.0
ply==3.11
pycparser==2.18
python-magic==0.4.15
pytz==2017.2
rcssmin==1.0.6
rjsmin==1.0.12
six==1.11.0
slimit==0.8.1
uWSGI==2.0.16
webencodings==0.5.1
wincertstore==0.2
